feat: Added comments to wg-init script to see the individual steps in the pod logs feat!: Repurposed tunneld image from standalone wireguard container.
The tunneld image now runs a cronjob which executes a py script. The py script gets the IPs
from all services which have a label, selected by the user, applied to it. For those IPs a
masquerade rule is created. With this the remote endpoint is able to reach the service ips in
the cluster.